Make response code more consistent.
