+++ functions/src/fulfillments/categoryFulfillment.ts
@@ -3,7 +3,7 @@ import { OptionQuestion, Question } from '../models/questions';
 
 import { Category } from '../models/categories';
 import { Topic } from '../models/rounds';
-import { buildSSMLAudioResponse } from '../responses/genericResponse';
+import { buildSSMLAudioResponse } from '../responses/ssmlResponses';
 import { categories } from '../content/categoryContent';
 import { unexpectedErrorResponse } from '../utils/logger';
